Seems to have memory leak, so trying to close everything on finally.
